[refac] Refaz com base em conselhos e ajudas de usuários da r/neovim
