+++ lua/ninote/init.lua
@@ -1,18 +1,25 @@
 local config = require("ninote.config")
 local ninote = {}
 
+local ARGUMENT_MODE = "+"
+local PLUGIN_NAME = "Ninote"
+local MODULE_LIST = {
+  "new",
+  "close",
+  "search",
+}
 
 function ninote.setup(user_config)
   config.load(user_config)
   vim.api.nvim_create_user_command(
-    "Ninote",
+    PLUGIN_NAME,
     function(args)
       require("ninote.router").route(args)
     end,
     {
-      nargs = "+",
+      nargs = ARGUMENT_MODE,
       complete = function()
-        return { "new", "close", "search" }
+        return MODULE_LIST
       end,
     }
   )
